Move URI/URL handling functions into separate module
