+++ rednotebook/gui/main_window.py
@@ -40,6 +40,7 @@ from rednotebook import templates
 from rednotebook.util import dates
 from rednotebook.util import markup
 from rednotebook.util import utils
+from rednotebook.util import urls
 from rednotebook.gui import categories
 from rednotebook.gui.exports import ExportAssistant
 from rednotebook.gui import browser
@@ -536,7 +537,8 @@ class MainWindow:
             if action.is_user_gesture():
                 uri = action.get_request().get_uri()
                 logging.info('Clicked URI "%s"' % uri)
-                filesystem.open_url(uri)
+
+                urls.open_url(uri)
 
                 decision.ignore()
